docs(storybook): add props tables for subcomponents
